garbage_project/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2
django-appconf==1.0.5
django-cities==0.6.1
django-crispy-forms==2.0
django-elasticsearch==7.0.2
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-model-utils==4.3.1
django-nine==0.2.7
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
flower==1.1.0
gunicorn==20.1.0
humanize==4.6.0
kombu==5.2.4
numpy==1.24.3
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
PyJWT==1.7.1
PyPrind==2.11.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
